Added Global Exceptions

+++ src/main/java/com/stackroute/service/TrackServiceImpl.java
@@ -1,6 +1,7 @@
 package com.stackroute.service;
 
 import com.stackroute.domain.Track;
+import com.stackroute.exceptions.TrackAlreadyExistsException;
 import com.stackroute.exceptions.TrackNotFoundException;
 import com.stackroute.repository.TrackRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,18 +11,22 @@ import java.util.List;
 
 @Service
 public class TrackServiceImpl implements TrackService {
-    TrackRepository trackRepository;
+    private TrackRepository trackRepository;
 
     public TrackServiceImpl(TrackRepository trackRepository) {
         this.trackRepository = trackRepository;
     }
 
     @Override
-    public Track saveTrack(Track track) {
-
+    public Track saveTrack(Track track) throws TrackAlreadyExistsException {
+        if (trackRepository.existsById(track.getId())) {
+            throw new TrackAlreadyExistsException("Track Already Exists");
+        }
         Track savedTrack = trackRepository.save(track);
+        if (savedTrack == null) {
+            throw new TrackAlreadyExistsException("Track Already Exists");
+        }
         return savedTrack;
-
     }
 
     @Override
@@ -31,18 +36,24 @@ public class TrackServiceImpl implements TrackService {
     }
 
     @Override
-    public Track updateTrackComments(Track track) {
+    public Track updateTrackComments(Track track) throws TrackNotFoundException {
+        if (!trackRepository.existsById(track.getId())) {
+            throw new TrackNotFoundException("Track Not Found");
+        }
         Track updatedTrack = trackRepository.save(track);
         return updatedTrack;
     }
 
     @Override
-    public void deleteTrack(int id) {
+    public void deleteTrack(int id) throws TrackNotFoundException {
+        if (!trackRepository.existsById(id)) {
+            throw new TrackNotFoundException("Track Not Found");
+        }
         trackRepository.deleteById(new Integer(id));
     }
-//
-//    @Override
-//    public Track getTrackById(int id) {
-//        return null;
-//    }
+
+    @Override
+    public List<Track> getByName(String name) {
+       return trackRepository.findByName(name);
+    }
 }
